arcade(invaders-mp): rename /invaders-2p/ → /invaders-mp/
Pure path/string rename ahead of the N-player work in subsequent
commits. No behaviour change — still 2-player gameplay, same wire
protocol, NETCODE_VERSION stays at 18.

- `docs/arcade/invaders-2p/` → `docs/arcade/invaders-mp/` (git mv,
  preserves blame)
- Server room.ts engine imports updated
- Client extractCode regex + cleanPath rewritten for the new path
- Page title + lobby header updated to "Invaders MP"
- arcade-site worker rewrites `/invaders-mp/<segment>` to the
  canonical index.html (same trick as before, just on the new path)
- arcade-site 301s `/invaders-2p/...` → `/invaders-mp/...` so any
  existing shared links / QRs from the previous spike still work
