+++ infra/oyster-arcade-site/src/worker.ts
@@ -22,16 +22,23 @@ export default {
       return fetch(new Request('https://oyster.to' + url.pathname + url.search, req));
     }
 
-    // Path-based room codes for the 2P spike — /invaders-2p/FROG (and
-    // /invaders-2p/FROG/ with trailing slash) should serve the
-    // canonical index.html so the client can read the room code from
-    // location.pathname. We rewrite any /invaders-2p/<segment> that
-    // doesn't look like a file (no dot) and isn't the bare directory
-    // path. Actual files (simple-peer.min.js, etc.) flow through to
-    // the ASSETS binding unchanged.
-    if (/^\/invaders-2p\/[^./]+\/?$/.test(url.pathname)) {
-      const rewritten = new URL('/invaders-2p/', req.url);
+    // Path-based room codes for the MP spike — /invaders-mp/FROG (and
+    // /invaders-mp/FROG/) should serve the canonical index.html so
+    // the client can read the room code from location.pathname. We
+    // rewrite any /invaders-mp/<segment> that doesn't look like a
+    // file (no dot) and isn't the bare directory path. Real files
+    // (simple-peer.min.js, engine.js, etc.) flow through to ASSETS.
+    if (/^\/invaders-mp\/[^./]+\/?$/.test(url.pathname)) {
+      const rewritten = new URL('/invaders-mp/', req.url);
       return env.ASSETS.fetch(new Request(rewritten, req));
+    }
+
+    // Back-compat for v18 (2P-only) shared links: 301 anything under
+    // /invaders-2p/ to the same path under /invaders-mp/. Preserves
+    // the room code so old QR codes / Messages links still work.
+    if (url.pathname === '/invaders-2p' || url.pathname.startsWith('/invaders-2p/')) {
+      const dest = '/invaders-mp' + url.pathname.slice('/invaders-2p'.length);
+      return Response.redirect(new URL(dest + url.search, req.url).toString(), 301);
     }
 
     return env.ASSETS.fetch(req);
